+++ taxonomy/api/v1/views.py
@@ -3,12 +3,16 @@ Taxonomy API views.
 """
 from rest_framework import permissions
 from rest_framework.mixins import ListModelMixin, RetrieveModelMixin
-from rest_framework.viewsets import GenericViewSet
+from rest_framework.viewsets import GenericViewSet, ModelViewSet
 
+from django_filters.rest_framework import DjangoFilterBackend
 from django.db.models import Prefetch
 
-from taxonomy.api.v1.serializers import JobPostingsSerializer, JobsListSerializer, SkillListSerializer
-from taxonomy.models import CourseSkills, Job, JobPostings, Skill
+from taxonomy.api.v1.serializers import (
+    JobPostingsSerializer, JobsListSerializer, SkillListSerializer, SkillsQuizSerializer
+)
+from taxonomy.models import CourseSkills, Job, JobPostings, Skill, SkillsQuiz
+from taxonomy.api.permissions import IsOwner
 
 
 class TaxonomyAPIViewSetMixin:
@@ -66,3 +70,32 @@ class JobPostingsViewSet(TaxonomyAPIViewSetMixin, RetrieveModelMixin, ListModelM
         return JobPostings.objects.all().select_related(
             'job'
         )
+
+
+class SkillsQuizViewSet(TaxonomyAPIViewSetMixin, ModelViewSet):
+    """
+    ViewSet to list and retrieve all JobPostings in the system.
+    """
+    serializer_class = SkillsQuizSerializer
+    permission_classes = (permissions.IsAuthenticated, IsOwner | permissions.IsAdminUser, )
+    filter_backends = (DjangoFilterBackend,)
+    filterset_fields = ('username', )
+
+    queryset = SkillsQuiz.objects.all()
+
+    def perform_create(self, serializer):
+        """
+        Attach the User to the SkillsQuiz Model by overriding perform_create method.
+        """
+        serializer.save(username=self.request.user)
+
+    def get_queryset(self):
+        """
+        User should only be able to access its own quiz.
+        """
+        queryset = self.queryset
+        # Non staff user should only be able to access their own records.
+        if not self.request.user.is_staff:
+            queryset = queryset.filter(username=self.request.user.username)
+
+        return queryset.all().select_related('current_job').prefetch_related('skills', 'future_jobs')